# public/api/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evaly API - Système de Gestion des Étudiants",
        "description": "API REST pour la gestion des étudiants, authentification et profils dans le système Evaly",
        "contact": {
            "email": "support@evaly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3002",
            "description": "Serveur API Evaly"
        }
    ],
    "paths": {
        "/api/students/login": {
            "post": {
                "tags": [
                    "Student Authentication"
                ],
                "summary": "Connexion étudiant",
                "description": "Authentifie un étudiant avec email et mot de passe",
                "operationId": "6e74d488f971bf9de4ac446920f650df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "etudiant@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "student": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur d'authentification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "requires_password_setup": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/setup-password": {
            "post": {
                "tags": [
                    "Student Authentication"
                ],
                "summary": "Définir le mot de passe initial",
                "description": "Permet à un étudiant de définir son mot de passe pour la première fois en utilisant son email",
                "operationId": "61508a790469ae11fb611743dfd228c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "motdepasse123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe défini avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "student": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "first_name": {
                                                            "type": "string",
                                                            "example": "Jean"
                                                        },
                                                        "last_name": {
                                                            "type": "string",
                                                            "example": "Dupont"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "jean.dupont@example.com"
                                                        },
                                                        "student_number": {
                                                            "type": "string",
                                                            "example": "STU202500001"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abcd...xyz"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mot de passe défini avec succès. Vous êtes maintenant connecté."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/change-password": {
            "post": {
                "tags": [
                    "Student Authentication"
                ],
                "summary": "Changer le mot de passe",
                "description": "Permet à un étudiant connecté de changer son mot de passe",
                "operationId": "ef91f136f0b44c7bc8ceb45171d6cf71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "ancienmdp123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "nouveaumdp123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "nouveaumdp123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/students/logout": {
            "post": {
                "tags": [
                    "Student Authentication"
                ],
                "summary": "Déconnexion étudiant",
                "description": "Déconnecte l'étudiant et révoque son token",
                "operationId": "b5af20f3e3f7291b346334771472a48d",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez le token Bearer obtenu lors de la connexion",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}